split list and autocomplete handlers
