+++ src/sparnatural/components/widgets/listwidget/LiteralListWidget.ts
@@ -1,16 +1,14 @@
-import { Pattern, ValuePatternRow, ValuesPattern } from "sparqljs";
+import { Pattern } from "sparqljs";
 import { SelectedVal } from "../../../generators/ISparJson";
-import { SparqlTemplateListHandler } from "../autocomplete/AutocompleteAndListHandlers";
 import WidgetWrapper from "../../builder-section/groupwrapper/criteriagroup/edit-components/WidgetWrapper";
-
 import * as DataFactory from "@rdfjs/data-model" ;
-import { Literal, Variable } from "@rdfjs/types";
-
+import { Literal } from "@rdfjs/types";
 import "select2";
 import "select2/dist/css/select2.css";
 import { ListWidget } from "./ListWidget";
-import { ValueRepetition, WidgetValue } from "../AbstractWidget";
+import { WidgetValue } from "../AbstractWidget";
 import SparqlFactory from "../../../generators/SparqlFactory";
+import { SparqlTemplateListHandler } from "./ListHandler";
 
 export class LiteralListWidgetValue implements WidgetValue {
   value: {
